move estimator to contrib (#14633)
